requirements.txt
Django==1.9
djangorestframework==3.3.2
Markdown==2.6.5
flake8==2.4.1
django-rest-swagger==0.3.5